telegram-bot-admin/requirements.txt
pip>=23.0

python-telegram-bot==13.15
grpcio
grpcio-tools
python-dotenv
pydantic
pydantic-settings
pandas
openpyxl

gspread
google-auth-oauthlib
oauth2client